add new model pvrcnn and voxelrcnn

+++ paddle3d/models/voxelizers/voxelize.py
@@ -41,7 +41,7 @@ class HardVoxelizer(nn.Layer):
             point, self.voxel_size, self.point_cloud_range,
             self.max_num_points_in_voxel, max_num_voxels)
         voxels = voxels[0:voxels_num, :, :]
-        coors = coors.astype(voxels.dtype)
+        #coors = coors.astype(voxels.dtype)
         coors = coors[0:voxels_num, :]
         num_points_per_voxel = num_points_per_voxel[0:voxels_num]
